ui_constructor in new app
