+++ tgbot/ui_constructor/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class UiConstructorConfig(AppConfig):
+    name = 'ui_constructor'
+    verbose_name = 'Конструктор бота'
